SSE connection problems Fixes #1
I've implemented a slight change in functionality where the API key is actually specified within the server and is not stipulated in the client request. This does limit the usability of this MCP server for multiple users, but for single user use it's absolutely fine. I would caution against any user who wants to actually use this server on a live web accessible server externally because there are other security implications. But I've added notes to that fact within the appropriate readme and Google Cloud hosting files. As is, this server now works properly for a specific API key user as most users are going to be running locally using Docker Compose or using Python directly on the command line. I think this is fine.